+++ tsconfig.json
@@ -22,7 +22,11 @@
     /* Additional Features */
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": true,
+
+    "paths": {
+      "@/*": ["src/*"]
+    }
   },
   "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.tsx", "src/**/*.vue", "src/vue-shims.d.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
